requirements.txt
astroid==1.6.5
beautifulsoup4==4.6.0
bs4==0.0.1
certifi==2018.4.16
chardet==3.0.4
coverage==3.7.1
coveralls==0.4.4
docopt==0.6.2
geojson==2.4.0
html5lib==1.0.1
idna==2.7
isort==4.3.4
lazy-object-proxy==1.3.1
lxml==4.2.3
mccabe==0.6.1
pycairo==1.17.1
pylint==1.9.2
python-dateutil==2.7.3
PyYAML==3.11
requests==2.4.3
six==1.11.0
urllib3==1.23
webencodings==0.5.1
wrapt==1.10.11
